Fix NPE on database migration
File#listFiles() may return null when the object it is applied to is not
a directory.

+++ main/src/cgeo/geocaching/cgData.java
@@ -677,10 +677,12 @@ public class cgData {
                     return filename.startsWith("map_") && filename.contains("__");
                 }
             };
-            for (File dir : geocodeDirs) {
-                File[] wrongFiles = dir.listFiles(filter);
-                for (File wrongFile : wrongFiles) {
-                    wrongFile.delete();
+            for (final File dir : geocodeDirs) {
+                final File[] wrongFiles = dir.listFiles(filter);
+                if (wrongFiles != null) {
+                    for (final File wrongFile : wrongFiles) {
+                        wrongFile.delete();
+                    }
                 }
             }
         }
